+++ src/components/search-bar/index.js
@@ -13,11 +13,7 @@ export const SearchBar = (props) => {
     props.setCanceled(true)
     props.setLoading(false)
     props.setInput(event.target.value)
-    if (event.target.value === "") {
-      props.setMessage("Search any GitHub username for a glimpse at their open source contributions")
-    } else {
-      props.setMessage(`Search for ${event.target.value}'s GitHub contributions`)
-    }
+    props.setMessage("")
   }
 
   const handleKeyPress = (event) => {